feat(packages/sui-precommit): Simplify precommit further
